refact: split container_stack resource code into multiple files
