+++ internal/provider/resource/containerstack/resource.go
@@ -2,7 +2,7 @@ package containerstackresource
 
 import (
 	"context"
-	"github.com/hashicorp/terraform-plugin-framework/diag"
+
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/mapplanmodifier"
@@ -11,14 +11,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-log/tflog"
 	mittwaldv2 "github.com/mittwald/api-client-go/mittwaldv2/generated/clients"
-	"github.com/mittwald/api-client-go/mittwaldv2/generated/clients/containerclientv2"
-	"github.com/mittwald/api-client-go/mittwaldv2/generated/schemas/containerv2"
-	"github.com/mittwald/terraform-provider-mittwald/internal/apiext"
 	"github.com/mittwald/terraform-provider-mittwald/internal/provider/providerutil"
 	"github.com/mittwald/terraform-provider-mittwald/internal/provider/resource/common"
-	"reflect"
 )
 
 var _ resource.Resource = &Resource{}
@@ -175,205 +170,6 @@ func (r *Resource) Schema(_ context.Context, _ resource.SchemaRequest, resp *res
 
 func (r *Resource) Configure(_ context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
 	r.client = providerutil.ClientFromProviderData(req.ProviderData, &resp.Diagnostics)
-}
-
-func (r *Resource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
-	var data ContainerStackModel
-
-	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	if data.DefaultStack.ValueBool() {
-		r.createInDefaultStack(ctx, data, resp)
-	} else {
-		r.createAsNewStack(ctx, data, resp)
-	}
-
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	resp.Diagnostics.Append(r.read(ctx, &data)...)
-	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
-}
-
-func (r *Resource) createAsNewStack(ctx context.Context, data ContainerStackModel, resp *resource.CreateResponse) {
-	client := apiext.NewContainerClient(r.client)
-
-	declareRequest := data.ToDeclareRequest(ctx, &resp.Diagnostics)
-	if resp.Diagnostics.HasError() {
-		tflog.Debug(ctx, "error while building declare request")
-		return
-	}
-
-	stack := providerutil.
-		Try[*containerv2.StackResponse](&resp.Diagnostics, "API error while declaring stack").
-		DoValResp(client.DeclareStack(ctx, *declareRequest))
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	data.ID = types.StringValue(stack.Id)
-}
-
-func (r *Resource) createInDefaultStack(ctx context.Context, data ContainerStackModel, resp *resource.CreateResponse) {
-	var current ContainerStackModel
-
-	client := apiext.NewContainerClient(r.client)
-
-	stack := providerutil.
-		Try[*containerv2.StackResponse](&resp.Diagnostics, "failed to get default stack").
-		DoVal(client.GetDefaultStack(ctx, data.ProjectID.ValueString()))
-
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	tflog.Debug(ctx, "using project default stack", map[string]any{"stack_id": stack.Id})
-
-	data.ID = types.StringValue(stack.Id)
-
-	updateRequest := data.ToUpdateRequest(ctx, &current, &resp.Diagnostics)
-	if resp.Diagnostics.HasError() {
-		tflog.Debug(ctx, "error while building update request")
-		return
-	}
-
-	_ = providerutil.
-		Try[*containerv2.StackResponse](&resp.Diagnostics, "API error while declaring stack").
-		DoValResp(r.client.Container().UpdateStack(ctx, *updateRequest))
-
-	return
-}
-
-func (r *Resource) read(ctx context.Context, data *ContainerStackModel) (res diag.Diagnostics) {
-	stack := providerutil.
-		Try[*containerv2.StackResponse](&res, "API error while fetching stack").
-		DoValResp(r.client.Container().GetStack(ctx, containerclientv2.GetStackRequest{StackID: data.ID.ValueString()}))
-
-	if res.HasError() {
-		return
-	}
-
-	res.Append(data.FromAPIModel(ctx, stack, true)...)
-
-	return
-}
-
-func (r *Resource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	data := ContainerStackModel{}
-
-	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	resp.Diagnostics.Append(r.read(ctx, &data)...)
-	resp.Diagnostics.Append(resp.State.Set(ctx, &data)...)
-}
-
-func (r *Resource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	var planData, stateData ContainerStackModel
-
-	resp.Diagnostics.Append(req.State.Get(ctx, &stateData)...)
-	resp.Diagnostics.Append(req.Plan.Get(ctx, &planData)...)
-
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	ctx = tflog.SetField(ctx, "stack_id", stateData.ID.ValueString())
-
-	var stack *containerv2.StackResponse
-
-	if stateData.DefaultStack.ValueBool() {
-		req := planData.ToUpdateRequest(ctx, &stateData, &resp.Diagnostics)
-		if resp.Diagnostics.HasError() {
-			return
-		}
-
-		stack = providerutil.
-			Try[*containerv2.StackResponse](&resp.Diagnostics, "API error while updating stack").
-			DoValResp(r.client.Container().UpdateStack(ctx, *req))
-	} else {
-		req := planData.ToDeclareRequest(ctx, &resp.Diagnostics)
-		if resp.Diagnostics.HasError() {
-			return
-		}
-
-		stack = providerutil.
-			Try[*containerv2.StackResponse](&resp.Diagnostics, "API error while declaring stack").
-			DoValResp(r.client.Container().DeclareStack(ctx, *req))
-	}
-
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	r.recreateContainers(ctx, planData, stack, resp)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	resp.Diagnostics.Append(r.read(ctx, &stateData)...)
-	resp.Diagnostics.Append(resp.State.Set(ctx, &stateData)...)
-}
-
-func (r *Resource) recreateContainers(ctx context.Context, planData ContainerStackModel, stack *containerv2.StackResponse, resp *resource.UpdateResponse) {
-	containerModels := planData.ContainerModels(ctx, &resp.Diagnostics)
-
-	for _, service := range stack.Services {
-		ctx := tflog.SetField(ctx, "service_id", service.Id)
-
-		serviceConfig, ok := containerModels[service.ServiceName]
-		if !ok {
-			continue
-		}
-
-		if reflect.DeepEqual(service.DeployedState, service.PendingState) {
-			tflog.Debug(ctx, "service has no pending state; skipping recreation")
-			continue
-		}
-
-		if serviceConfig.NoRecreateOnChange.ValueBool() {
-			tflog.Debug(ctx, "recreation would be necessary, but no_recreate_on_change is set; skipping recreation")
-			continue
-		}
-
-		req := containerclientv2.RecreateServiceRequest{
-			StackID:   stack.Id,
-			ServiceID: service.Id,
-		}
-
-		tflog.Debug(ctx, "recreating service")
-
-		providerutil.
-			Try[any](&resp.Diagnostics, "API error while recreating container").
-			DoResp(r.client.Container().RecreateService(ctx, req))
-	}
-}
-
-func (r *Resource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
-	var stateData ContainerStackModel
-
-	resp.Diagnostics.Append(req.State.Get(ctx, &stateData)...)
-
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	// The "default" stack has a special role, and we will not delete it even if
-	// the user requests it. Instead, we will simply purge all containers and volumes
-	// from it
-	if stateData.DefaultStack.ValueBool() {
-		_ = providerutil.
-			Try[*containerv2.StackResponse](&resp.Diagnostics, "API error while removing containers from stack").
-			DoValResp(r.client.Container().UpdateStack(ctx, *stateData.ToDeletePatchRequest(ctx, &resp.Diagnostics)))
-	} else {
-		resp.Diagnostics.AddError("not implemented", "removing non-default stacks is not supported, yet")
-	}
 }
 
 func (r *Resource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
